pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcore-app"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
    "bedrock-agentcore>=1.2.1",
    "pydantic>=2.0.0",
    "strands-agents>=1.25.0",
    "strands-agents-tools>=0.2.20",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.40.0",
    "browser-use>=0.1.0",
    "langchain-aws>=0.2.0",
]
all = [
    "playwright>=1.40.0",
    "browser-use>=0.1.0",
    "langchain-aws>=0.2.0",
]

[dependency-groups]
dev = [
    "bedrock-agentcore-starter-toolkit>=0.3.2",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
